+++ backend/src/main/kotlin/org/setackle/backend/adapter/inbound/web/user/controller/AuthController.kt
@@ -6,7 +6,11 @@ import jakarta.servlet.http.HttpServletRequest
 import jakarta.validation.Valid
 import org.setackle.backend.adapter.inbound.web.common.ApiResponse
 import org.setackle.backend.adapter.inbound.web.user.dto.*
-import org.setackle.backend.domain.user.inbound.*
+import org.setackle.backend.application.user.inbound.LoginUseCase
+import org.setackle.backend.application.user.inbound.LogoutAllSessionsUseCase
+import org.setackle.backend.application.user.inbound.LogoutUseCase
+import org.setackle.backend.application.user.inbound.RefreshTokenUseCase
+import org.setackle.backend.application.user.inbound.RegisterUserUseCase
 import org.springframework.http.HttpStatus
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
