+++ backend/src/main/kotlin/org/setackle/backend/adapter/outbound/persistence/security/CustomUserDetailsService.kt
@@ -1,7 +1,7 @@
 package org.setackle.backend.adapter.outbound.persistence.security
 
-import org.setackle.backend.application.security.CustomUserDetails
 import org.setackle.backend.adapter.outbound.persistence.user.repository.UserJpaRepository
+import org.setackle.backend.application.security.CustomUserDetails
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UserDetailsService
 import org.springframework.security.core.userdetails.UsernameNotFoundException
