+++ backend/src/main/kotlin/org/setackle/backend/adapter/outbound/security/TokenAdapter.kt
@@ -2,7 +2,7 @@ package org.setackle.backend.adapter.outbound.security
 
 import org.setackle.backend.application.security.CustomUserDetails
 import org.setackle.backend.domain.user.model.User
-import org.setackle.backend.domain.user.outbound.TokenPort
+import org.setackle.backend.application.user.outbound.TokenPort
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.Authentication
 import org.springframework.stereotype.Component
