+++ backend/src/main/kotlin/org/setackle/backend/application/user/inbound/LogoutUseCase.kt
@@ -1,4 +1,4 @@
-package org.setackle.backend.domain.user.inbound
+package org.setackle.backend.application.user.inbound
 
 /**
  * 로그아웃 유스케이스
