+++ backend/src/main/kotlin/org/setackle/backend/domain/user/model/User.kt
@@ -1,5 +1,10 @@
 package org.setackle.backend.domain.user.model
 
+import org.setackle.backend.domain.user.vo.Email
+import org.setackle.backend.domain.user.vo.Password
+import org.setackle.backend.domain.user.vo.UserId
+import org.setackle.backend.domain.user.vo.UserRole
+import org.setackle.backend.domain.user.vo.Username
 import java.time.LocalDateTime
 
 /**
