+++ backend/src/main/kotlin/org/setackle/backend/domain/user/vo/SessionId.kt
@@ -1,4 +1,4 @@
-package org.setackle.backend.domain.user.model
+package org.setackle.backend.domain.user.vo
 
 import java.util.*
