+++ backend/src/main/kotlin/org/setackle/backend/domain/user/vo/TokenPair.kt
@@ -1,4 +1,4 @@
-package org.setackle.backend.domain.user.model
+package org.setackle.backend.domain.user.vo
 
 /**
  * Token 쌍 값 객체
