+++ backend/src/main/kotlin/org/setackle/backend/domain/user/vo/UserId.kt
@@ -1,4 +1,4 @@
-package org.setackle.backend.domain.user.model
+package org.setackle.backend.domain.user.vo
 
 /**
  * 사용자 ID 값 객체
