+++ backend/src/main/kotlin/org/setackle/backend/domain/user/vo/UserRole.kt
@@ -1,4 +1,4 @@
-package org.setackle.backend.domain.user.model
+package org.setackle.backend.domain.user.vo
 
 enum class UserRole(val displayName: String) {
     USER("일반 사용자"),
